FETCH FOR ME 2

+++ veterinaryWebServices/src/main/java/com/team12/veterinaryWebServices/viewmodel/petVM.java
@@ -2,7 +2,7 @@ package com.team12.veterinaryWebServices.viewmodel;
 
 import com.team12.veterinaryWebServices.model.pet;
 
-public record petVM (Long petID, String petNAME, String imageLink, String petSPECIE, Long petAGE){
+public record petVM (Long petID, String petNAME, String imageLink, String petSPECIE, int petAGE){
     public static petVM from (pet p){
 
         return new petVM(
